refactor: exceptions moved to other package to organize with the implementation in BOOT-811
