+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/domains/exceptions/tokens/BadAuthRequestException.java
@@ -1,4 +1,4 @@
-package com.ciandt.summit.bootcamp2022.domains.token.exceptions;
+package com.ciandt.summit.bootcamp2022.domains.exceptions.tokens;
 
 public class BadAuthRequestException extends RuntimeException {
     public BadAuthRequestException(String message) {
