+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/domains/exceptions/tokens/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package com.ciandt.summit.bootcamp2022.domains.token.exceptions;
+package com.ciandt.summit.bootcamp2022.domains.exceptions.tokens;
 
 public class UnauthorizedException extends RuntimeException {
     public UnauthorizedException(String message) {
